requirements.txt
monzo==0.10.0
python-dateutil==2.8.1
requests==2.20.0